+++ packages/extensions/openapi-to-typespec/test/arm-alertsmanagement/tsp-output/routes.tsp
@@ -14,7 +14,7 @@ using TypeSpec.OpenAPI;
 namespace Azure.ResourceManager.AlertsManagement;
 
 #suppress "@azure-tools/typespec-azure-resource-manager/arm-resource-interface-requires-decorator" "For backward compatibility"
-interface AlertsOperations {
+interface AlertsOperationGroup {
   /**
    * List alerts meta data information based on value of identifier parameter.
    */
